honksquad: refactor SkillchipStationSystem into three subsystems (#859)
Split the 397-line SkillchipStationSystem into its three loosely-coupled
concerns, a behavior-preserving extraction:

- SkillchipStationOccupantManager owns the body-container occupant state
  machine: drag-drop boarding, relay-move / destruction eject, and the
  occupied/open appearance. Exposes IsOccupied / TryGetOccupant.
- SkillchipStationUiBuilder owns UI-state serialization and its refresh
  triggers (UI opened, container changes), consolidating the chip-slot
  and occupant lookups in one place.
- SkillchipStationSystem keeps the chip-tray interaction and the powered
  implant/remove flow, orchestrating the two helpers.

A shared TryResolveOperationBrain helper kills the occupant-lookup +
validate duplication between the implant and remove do-afters, and
TryStartOperation folds their identical do-after-start tail. The
install flow keeps its missing-occupant / missing-brain popups; the
remove flow stays silent as before.

Covered by SkillchipGrantTest.cs and SkillchipStationTest.cs.
